modified styling of spotlight_container and navbar to have horizontal margins when on smaller screens. moved JSON-generated pages to new generated blueprint. Renamed blueprints to consistent internal name <name>_bp
