+++ main.py
@@ -1,19 +1,20 @@
 from flask import Flask, render_template, send_from_directory, request
 import json
 import os
-import glob
 import math
 from datetime import datetime
 from werkzeug.exceptions import HTTPException
 from util import validate_json_schema, get_json_file_based_on_date
-from errors import raise_joke_http_error, JokeError
-from renders import render_validated_json_template, render_http_error, render_recipe_page, render_human_validator
-from fixed import fixed_bp
+from errors import JokeError
+from renders import render_http_error
+from bp_fixed import fixed_bp
+from bp_generated import generated_bp
 
 
 app = Flask(__name__)
 app.template_folder = 'templates'
 app.register_blueprint(fixed_bp)
+app.register_blueprint(generated_bp)
 boot_date = datetime.now()
 
 # create global config data on boot
@@ -58,12 +59,6 @@ def inject_global_data():
     else:
         raise FileNotFoundError("Invalid Schema")
 
-# open and load recipe and cardlist schema to avoid loading them every time either is needed
-recipes_schema_path = os.path.join(app.static_folder, 'dist', 'src', 'json', 'schema', 'recipes.json')
-cardlist_schema_path = os.path.join(app.static_folder, 'dist', 'src', 'json', 'schema', 'cardlist.json')
-with open(recipes_schema_path) as r, open(cardlist_schema_path) as c:
-        recipe_schema = json.load(r)
-        cardlist_schema = json.load(c)
 
 @app.route('/')
 def index():
@@ -72,70 +67,6 @@ def index():
 @app.route('/static/<path:filename>')
 def static_files(filename):
     return send_from_directory('static', filename)
-
-@app.route('/recipes')
-def recipes():
-    recipes_folder_path = os.path.join(app.static_folder, 'dist', 'src', 'json', 'recipes')
-
-    # get query from request, otherwise it is ""
-    search_query = request.args.get("query", "").lower()
-    selected_meals = request.args.getlist("meal")
-
-    recipes_data = []
-    invalid_file_cnt = 0
-    recipe_files = glob.glob(os.path.join(recipes_folder_path, '*.json')) # get full paths to all JSON files in the recipes folder
-    for recipe_file in recipe_files:
-        recipe_link = os.path.splitext(os.path.basename(recipe_file))[0] # link used within site
-        with open(recipe_file) as f:
-            recipe = json.load(f)
-            if validate_json_schema(recipe, recipe_schema): # if recipe matches schema, build data for use on page
-                title = recipe.get("title", "").lower()
-                recipe_meals = recipe.get("meal_type", [])
-                if isinstance(recipe_meals, str):
-                    recipe_meals = [recipe_meals.lower()]
-                else:
-                    recipe_meals = [m.lower() for m in recipe_meals]
-
-                title_match = search_query in title if search_query else True
-                meals_match = all(meal in recipe_meals for meal in selected_meals) if selected_meals else True
-
-                if title_match and meals_match:
-                    recipes_data.append(
-                        {"link": f"recipes/{recipe_link}", "data": recipe}
-                        )
-            else:
-                invalid_file_cnt += 1
-                print(f"❌ {f}: Invalid file schema")
-
-    return render_template('/sections/recipes/browse.html', recipes=recipes_data, invalid_count=invalid_file_cnt)
-
-@app.route('/recipe_substitutions')
-def recipe_substitutions():
-    substitutions_file = os.path.join(app.static_folder, 'dist', 'src', 'json', 'static', "recipe_substitutions.json")
-    return render_validated_json_template('/sections/general/index_card_list.html', substitutions_file, cardlist_schema)
-
-@app.route('/recipes/<path:recipe_name>')
-def recipe_detail(recipe_name):
-    return render_recipe_page(recipe_name, recipe_schema)
-
-@app.route('/recipes/review/<path:recipe_name>')
-def recipe_review(recipe_name):
-    return render_recipe_page(recipe_name, recipe_schema, True)
-
-@app.route('/recipes/submitreview/<path:recipe_name>')
-def submit_recipe_review(recipe_name):
-    return raise_joke_http_error("advice_ignored", app.global_config["ERRORS"]["custom_codes"]["advice_ignored"]["description"])
-
-@app.route('/sex_positions')
-def sex_positions():
-    sex_positions_file = os.path.join(app.static_folder, 'dist', 'src', 'json', 'static', 'sex_positions.json')
-
-    cookie = request.cookies.get('sex_pos_validator')
-    print(f"COOKIE : {cookie}")
-    if cookie == "T":
-        return render_validated_json_template('/sections/general/index_card_list.html', sex_positions_file, cardlist_schema)
-    else:
-        return render_human_validator('sex_pos_validator', request.path)
 
 @app.errorhandler(JokeError)
 def handle_joke_error(e):
