+++ templates/sections/general/index.html
@@ -28,7 +28,7 @@
             {% if RECIPE_OF_THE_DAY is none %}
                 <p>None</p>
             {% else %}
-                <a href="{{ url_for('recipe_detail', recipe_name=RECIPE_OF_THE_DAY.link) }}">{{ RECIPE_OF_THE_DAY.display_text }}</a>
+                <a href="{{ url_for('generated_bp.recipe_detail', recipe_name=RECIPE_OF_THE_DAY.link) }}">{{ RECIPE_OF_THE_DAY.display_text }}</a>
             {% endif %}
         </div>
     </div>
