+++ templates/sections/recipes/browse.html
@@ -4,7 +4,7 @@
 
 {% block content %}
     <div class="flex flex-row mx-2 lg:mx-auto gap-4 items-center">
-        <form class="widget_container p-1 px-2" method="get" action="{{ url_for('recipes') }}">
+        <form class="widget_container p-1 px-2" method="get" action="{{ url_for('generated_bp.recipes') }}">
             <input type="text" name="query" placeholder="Search title..." value="{{ request.args.get('query', '') }}">
         
             <label><input type="checkbox" name="meal" value="breakfast"
