+++ templates/sections/recipes/view.html
@@ -29,10 +29,10 @@
     {% if not list %}
         N/A
     {% elif list is mapping %}
-        <a class="info_content" href="{{ url_for('recipe_detail', recipe_name=list.json_filename) }}">{{ list.displaytext }}</a>
+        <a class="info_content" href="{{ url_for('generated_bp.recipe_detail', recipe_name=list.json_filename) }}">{{ list.displaytext }}</a>
     {% else %}
         {% for item in list %}
-            <a class="info_content" href="{{ url_for('recipe_detail', recipe_name=item.json_filename) }}">{{ item.displaytext }}{% if not loop.last %},{% endif %}</a>
+            <a class="info_content" href="{{ url_for('generated_bp.recipe_detail', recipe_name=item.json_filename) }}">{{ item.displaytext }}{% if not loop.last %},{% endif %}</a>
         {% endfor %}
     {% endif %}
 {% endmacro %}
@@ -227,6 +227,6 @@
         </div>
     </div>
     <div class="flex flex-1 flex-row items-center justify-center pt-6 pb-4">
-        <a class="buttonlink" href="{{ url_for('recipe_review', recipe_name=additional_data.recipe_name) }}">Rate this Recipe!</a>
+        <a class="buttonlink" href="{{ url_for('generated_bp.recipe_review', recipe_name=additional_data.recipe_name) }}">Rate this Recipe!</a>
     </div>
 {% endblock %}
